+++ node/clients/index.ts
@@ -1,8 +1,13 @@
 import { IOClients } from '@vtex/api'
 
+import { CallCenterOperator } from './callCenterOperator'
+import { Catalog } from './catalog'
 import { Checkout } from './checkout'
 import { MasterData } from './masterdata'
-import { CallCenterOperator } from './callCenterOperator'
+import { ProfileClient } from './profile'
+import { OMS } from './oms'
+import { Portal } from './portal'
+import { LogisticsClient } from './logistics'
 
 export class Clients extends IOClients {
   public get masterdata() {
@@ -15,5 +20,25 @@ export class Clients extends IOClients {
 
   public get callCenterOperator() {
     return this.getOrSet('callCenterOperator', CallCenterOperator)
+  }
+
+  public get catalog() {
+    return this.getOrSet('catalog', Catalog)
+  }
+
+  public get profile() {
+    return this.getOrSet('profile', ProfileClient)
+  }
+
+  public get oms() {
+    return this.getOrSet('oms', OMS)
+  }
+
+  public get portal() {
+    return this.getOrSet('portal', Portal)
+  }
+
+  public get logistics() {
+    return this.getOrSet('logistics', LogisticsClient)
   }
 }