+++ node/resolvers/catalog/discount.ts
@@ -3,5 +3,8 @@ import { prop } from 'ramda'
 export const resolvers = {
   Discount: {
     name: prop('<Name>k__BackingField')
+  },
+  Teaser: {
+    name: prop('<Name>k__BackingField')
   }
 }